+++ src/plauna/markup.clj
@@ -238,11 +238,18 @@
          :started (some-> (:started-at batch) timestamp->date str (str/replace "T" " ") (subs 0 16))
          :finished (some-> (:finished-at batch) timestamp->date str (str/replace "T" " ") (subs 0 16))))
 
+(defn- format-timestamp [epoch-seconds]
+  (some-> epoch-seconds timestamp->date str (str/replace "T" " ") (subs 0 16)))
+
+(defn- format-parse-failure-group [group]
+  (update group :failures (fn [failures] (mapv #(assoc % :last-seen-text (format-timestamp (:last-seen %))) failures))))
+
 (defn- connection-context [config folders categories]
   (merge config {:folders folders
                  :active-nav :admin
                  :categories (cons nil categories)
-                 :parse-batches (mapv format-parse-batch (:parse-batches config))}))
+                 :parse-batches (mapv format-parse-batch (:parse-batches config))
+                 :parse-failures (mapv format-parse-failure-group (:parse-failures config))}))
 
 (defn connection
   ([config folders categories] (render "admin-connection.html" (connection-context config folders categories)))